Improved documentation Refactored base `TileProvider` into independent file
