+++ lib/src/layer/tile_layer/tile_provider/tile_provider_web.dart
@@ -4,63 +4,6 @@ import 'package:http/retry.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_provider/network_image_provider.dart';
 
-abstract class TileProvider {
-  Map<String, String> headers;
-
-  TileProvider({
-    this.headers = const {},
-  });
-
-  ImageProvider getImage(Coords coords, TileLayerOptions options);
-
-  void dispose() {}
-
-  String getTileUrl(Coords coords, TileLayerOptions options) {
-    final urlTemplate = (options.wmsOptions != null)
-        ? options.wmsOptions!
-            .getUrl(coords, options.tileSize.toInt(), options.retinaMode)
-        : options.urlTemplate;
-
-    final z = _getZoomForUrl(coords, options);
-
-    final data = <String, String>{
-      'x': coords.x.round().toString(),
-      'y': coords.y.round().toString(),
-      'z': z.round().toString(),
-      's': getSubdomain(coords, options),
-      'r': '@2x',
-    };
-    if (options.tms) {
-      data['y'] = invertY(coords.y.round(), z.round()).toString();
-    }
-    final allOpts = Map<String, String>.from(data)
-      ..addAll(options.additionalOptions);
-    return options.templateFunction(urlTemplate!, allOpts);
-  }
-
-  double _getZoomForUrl(Coords coords, TileLayerOptions options) {
-    var zoom = coords.z;
-
-    if (options.zoomReverse) {
-      zoom = options.maxZoom - zoom;
-    }
-
-    return zoom += options.zoomOffset;
-  }
-
-  int invertY(int y, int z) {
-    return ((1 << z) - 1) - y;
-  }
-
-  String getSubdomain(Coords coords, TileLayerOptions options) {
-    if (options.subdomains.isEmpty) {
-      return '';
-    }
-    final index = (coords.x + coords.y).round() % options.subdomains.length;
-    return options.subdomains[index];
-  }
-}
-
 /// [TileProvider] that uses [FMNetworkImageProvider] internally
 ///
 /// This image provider automatically retries some failed requests up to 3 times.
@@ -131,6 +74,9 @@ class AssetTileProvider extends TileProvider {
   }
 }
 
+/// A very basic [TileProvider] implementation, that can be extended to create your own provider
+///
+/// Using this method is not recommended any more, except for very simple custom [TileProvider]s. Instead, visit the online documentation at https://docs.fleaflet.dev/plugins/making-a-plugin/creating-new-tile-providers.
 class CustomTileProvider extends TileProvider {
   final String Function(Coords coors, TileLayerOptions options) customTileUrl;
